+++ src/pages-components/Dashboard/ClientsPage/SingleClientPage/SingleClientPage.tsx
@@ -1,10 +1,10 @@
 import { useRouter } from 'next/router';
 
+import { BackButton } from '@/components/BackButton';
 import { DashboardLayout } from '@/components/layout/DashboardLayout';
 import type { Client } from '@/lib/api';
 import { LoadingPage } from '@/lib/components/LoadingPage';
 import { useClient } from '@/lib/hooks';
-import { BackButton } from '@/pages-components/Dashboard/NewEmployeePage/BackButton';
 import { getCountryFromCode } from '@/pages-components/Dashboard/NewEmployeePage/const';
 import { Box, Divider, FlexLayout, Text } from '@/ui';
