+++ src/pages-components/Dashboard/ContractorsPage/EditContractorPage/EditContractorPage.tsx
@@ -1,11 +1,11 @@
 import { useRouter } from 'next/router';
 
+import { BackButton } from '@/components/BackButton';
 import { DashboardLayout } from '@/components/layout/DashboardLayout';
 import type { Contractor } from '@/lib/api';
 import { LoadingPage } from '@/lib/components/LoadingPage';
 import { useContractor } from '@/lib/hooks';
 import { NewContractorForm } from '@/pages-components/Dashboard/ContractorsPage/NewContractorPage/NewContractorForm';
-import { BackButton } from '@/pages-components/Dashboard/NewEmployeePage/BackButton';
 import { Box, FlexLayout, Heading } from '@/ui';
 
 export const EditContractorPage = () => {
