+++ src/pages-components/Dashboard/ContractorsPage/NewContractorPage/NewContractorPage.tsx
@@ -1,5 +1,5 @@
+import { BackButton } from '@/components/BackButton';
 import { DashboardLayout } from '@/components/layout/DashboardLayout';
-import { BackButton } from '@/pages-components/Dashboard/NewEmployeePage/BackButton';
 import { Box, FlexLayout, Heading } from '@/ui';
 
 import { NewContractorForm } from './NewContractorForm';
