+++ src/pages-components/Dashboard/ContractorsPage/SingleContractorPage/SingleContractorPage.tsx
@@ -1,10 +1,10 @@
 import { useRouter } from 'next/router';
 
+import { BackButton } from '@/components/BackButton';
 import { DashboardLayout } from '@/components/layout/DashboardLayout';
 import type { Contractor } from '@/lib/api';
 import { LoadingPage } from '@/lib/components/LoadingPage';
 import { useContractor } from '@/lib/hooks';
-import { BackButton } from '@/pages-components/Dashboard/NewEmployeePage/BackButton';
 import { getCountryFromCode } from '@/pages-components/Dashboard/NewEmployeePage/const';
 import { Box, Divider, FlexLayout, Text } from '@/ui';
