+++ src/pages-components/Dashboard/EditEmployeePage/EditEmployeePage.tsx
@@ -1,12 +1,12 @@
 import { useRouter } from 'next/router';
 
+import { BackButton } from '@/components/BackButton';
 import { DashboardLayout } from '@/components/layout/DashboardLayout';
 import type { Employee } from '@/lib/api';
 import { LoadingPage } from '@/lib/components/LoadingPage';
 import { useEmployee } from '@/lib/hooks';
 import { Box, FlexLayout, Heading } from '@/ui';
 
-import { BackButton } from '../NewEmployeePage/BackButton';
 import { NewEmployeeForm } from '../NewEmployeePage/NewEmployeeForm';
 
 export const EditEmployeePage = () => {
