+++ src/pages-components/Dashboard/FleetPage/Trailers/SingleTrailerPage/SingleTrailerPage.tsx
@@ -1,10 +1,10 @@
 import { useRouter } from 'next/router';
 
+import { BackButton } from '@/components/BackButton';
 import { DashboardLayout } from '@/components/layout/DashboardLayout';
 import type { Vehicle } from '@/lib/api';
 import { LoadingPage } from '@/lib/components/LoadingPage';
 import { useVehicle } from '@/lib/hooks';
-import { BackButton } from '@/pages-components/Dashboard/NewEmployeePage/BackButton';
 import { Box, FlexLayout, Icon, Pill, Text } from '@/ui';
 
 import { GeneralInfo } from '../../components/GeneralInfo';
